% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.r
\name{get_available_aggregations}
\alias{get_available_aggregations}
\title{Get available aggregations}
\usage{
get_available_aggregations(r)
}
\arguments{
\item{r}{The results list
Should be of length 3 with elements: \code{params}, \code{population_variants} and \code{results}}
}
\value{
A list of the possible aggregations
}
\description{
Get available aggregations
}
\examples{
filename <- system.file("sample_results.json", package = "nhp.results")
results <- get_results_from_local(filename)
get_trust_sites(results)
}
