% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.r
\name{get_model_run_years}
\alias{get_model_run_years}
\title{Get model_run_years}
\usage{
get_model_run_years(r)
}
\arguments{
\item{r}{The results list
Should be of length 3 with elements: \code{params}, \code{population_variants} and \code{results}}
}
\value{
A list of 2, containing \code{start_year} and \code{end_year}
}
\description{
Get model_run_years
}
\examples{
filename <- system.file("sample_results.json", package = "nhp.results")
results <- get_results_from_local(filename)
get_model_run_years(results)
}
