% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.r
\name{get_results_from_local}
\alias{get_results_from_local}
\title{Get results from local}
\usage{
get_results_from_local(filename)
}
\arguments{
\item{filename}{Filename for model results JSON}
}
\description{
Get results from local
}
\examples{
filename <- system.file("sample_results.json", package = "nhp.results")
results <- get_results_from_local(filename)
}
