% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.r
\name{get_trust_sites}
\alias{get_trust_sites}
\title{Get trust sites}
\usage{
get_trust_sites(r)
}
\arguments{
\item{r}{The results list
Should be of length 3 with elements: \code{params}, \code{population_variants} and \code{results}}
}
\value{
A vector of site names
}
\description{
Get trust sites
}
\examples{
filename <- system.file("sample_results.json", package = "nhp.results")
results <- get_results_from_local(filename)
get_trust_sites(results)
}
