% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.r
\name{get_variants}
\alias{get_variants}
\title{Get variants
Get the population variants}
\usage{
get_variants(r)
}
\arguments{
\item{r}{The results list
Should be of length 3 with elements: \code{params}, \code{population_variants} and \code{results}}
}
\value{
A tibble with two columns, \code{model_run} & \code{variant}
}
\description{
Get variants
Get the population variants
}
\examples{
filename <- system.file("sample_results.json", package = "nhp.results")
results <- get_results_from_local(filename)
get_variants(results)
}
