% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.r
\name{patch_results}
\alias{patch_results}
\title{Patch results}
\usage{
patch_results(r)
}
\arguments{
\item{r}{The results list
Should be of length 3 with elements: \code{params}, \code{population_variants} and \code{results}}
}
\description{
Patch results
}
